refactor: delete console.log
